Revert to old behavior
It seems that after adding `hidden yes` to the command, we're getting more instances
of the car saying that your credintials are incorrect. This didn't use to happen.
Fortunatly the times it says your credentials are wrong, if you just try again with
the same credentials it would start working (assuming that they were, indeed, correct).
Anyway, reverting anyway.