Increase timeout when checking for internet access
The theory is that the getaddrinfo was very slow due to checking for AAAA
records. See https://unix.stackexchange.com/a/141174

I saw evidence of this in testing at one school. I saw that getaddrinfo would
take either 5, 10, or 15 seconds to return (that's why the timeout is now 60
seconds). Weird that it would always take some mulitple of 5 seconds. That,
to me, indicates that a DNS server somewhere has a 5 second timeout, and
perhaps there is sometimes a chaining effect to this from one DNS to the
next? Anyway, I futher saw evidence that it seems to be an IPv6 AAAA record
issue that causes the delay. When I specified `family=socket.AF_INET` to
the `getaddrinfo` function, it would return immediately!

All told, it feels that a safe solution is to increase the timeouts all
around. So, that's what has been done.