Don't switch users anymore
This was cleaver, but it was not a good idea really. Especially now
with all the concerancy that is going on, switching the prcesse's
user and group can cause issues with the other work that is happening
in the other asyncio tasks (or just simply in the other threads if
any other threads are running). E.g. I was seeing a deadlock in the
camera driver and traced it down to the time of the user switching.
I think the kernel got confused when we had the camera open and
then tried to switch users. Not sure really, but I know removing this
fixed the deadlock issue, and generally I think it's best we don't
have our process switch users and groups in this highly concurrent
world we have now.