Reverting
Change of plans from 187d0e29f38350d1a6f877fa71cec0ed56309c94.

Jupyter's terminal starts in the directory where you run the `jupyter notebook` command. It's important that we have the terminal open to the home directory.

We'll have to revisit this for a better fix at a later date.